feat: address page support `node-connect` mode
