+++ src/locales/zh.json
@@ -710,7 +710,9 @@
       "current_page": "第",
       "of_page": "页，共",
       "only_first_pages_visible": "显示最近 {{pages}} 页",
-      "show_rows": "每页显示"
+      "show_rows": "每页显示",
+      "load_more": "加载更多",
+      "no_more_data": "没有更多数据了"
     },
     "udt": {
       "sudt": "sUDT",
