+++ src/pages/Transaction/TransactionCell/NodeTransactionCell.tsx
@@ -2,11 +2,9 @@ import { useState, ReactNode } from 'react'
 import { Tooltip } from 'antd'
 import type { Cell } from '@ckb-lumos/base'
 import { useTranslation } from 'react-i18next'
-import { useQuery } from '@tanstack/react-query'
 import { Link } from '../../../components/Link'
 import { IOType } from '../../../constants/common'
-import { parseUDTAmount } from '../../../utils/number'
-import { shannonToCkb, shannonToCkbDecimal } from '../../../utils/util'
+import { shannonToCkbDecimal } from '../../../utils/util'
 import {
   TransactionCellContentPanel,
   TransactionCellDetailPanel,
@@ -20,7 +18,7 @@ import {
   TransactionCellMobileItem,
 } from './styled'
 import { LeftArrow } from '../../../components/Transaction/TransactionCellArrow'
-import Capacity from '../../../components/Capacity'
+import { NodeCellCapacityAmount } from '../../../components/TransactionItem/TransactionItemCell/NodeCellCapacityAmount'
 import NervosDAODepositIcon from '../../../assets/nervos_dao_cell.png'
 import NervosDAOWithdrawingIcon from '../../../assets/nervos_dao_withdrawing.png'
 import UDTTokenIcon from '../../../assets/udt_token.png'
@@ -34,8 +32,7 @@ import { CellInfoModal } from '../TransactionCellScript'
 import SimpleModal from '../../../components/Modal'
 import SimpleButton from '../../../components/SimpleButton'
 import { useIsMobile } from '../../../hooks'
-import { explorerService } from '../../../services/ExplorerService'
-import { UDT_CELL_TYPES, getCellType, calculateScriptHash, getUDTAmountByData } from '../../../utils/cell'
+import { getCellType } from '../../../utils/cell'
 import { encodeNewAddress, encodeDeprecatedAddress } from '../../../utils/address'
 import { Addr } from './index'
 
@@ -195,39 +192,6 @@ const TransactionCellInfo = ({
   )
 }
 
-const TransactionCellCapacityAmount = ({ cell }: { cell: Cell }) => {
-  const { t } = useTranslation()
-  const cellType = getCellType(cell)
-  const isUDTCell = UDT_CELL_TYPES.findIndex(type => type === cellType) !== -1
-  const udtTypeHash = isUDTCell ? calculateScriptHash(cell.cellOutput.type!) : undefined
-  const udtInfo = useQuery(
-    ['udt', udtTypeHash],
-    () => {
-      if (!udtTypeHash) return undefined
-      return explorerService.api.fetchSimpleUDT(udtTypeHash)
-    },
-    {
-      enabled: isUDTCell,
-      staleTime: Infinity,
-    },
-  )
-
-  if (isUDTCell && udtTypeHash && udtInfo.data) {
-    const amount = getUDTAmountByData(cell.data)
-    if (cellType === 'udt' && udtInfo.data.published) {
-      return <span>{`${parseUDTAmount(amount, udtInfo.data.decimal)} ${udtInfo.data.symbol}`}</span>
-    }
-
-    if (cellType === 'xudt' && udtInfo.data.decimal && udtInfo.data.symbol) {
-      return <span>{`${parseUDTAmount(amount, udtInfo.data.decimal)} ${udtInfo.data.symbol}`}</span>
-    }
-
-    return <span>{`${t('udt.unknown_token')} #${udtTypeHash.substring(udtTypeHash.length - 4)}`}</span>
-  }
-
-  return <Capacity capacity={shannonToCkb(cell.cellOutput.capacity)} layout="responsive" />
-}
-
 export default ({
   cell,
   index,
@@ -257,10 +221,7 @@ export default ({
             </TransactionCellInfo>
           }
         />
-        <TransactionCellMobileItem
-          title={t('transaction.capacity')}
-          value={<TransactionCellCapacityAmount cell={cell} />}
-        />
+        <TransactionCellMobileItem title={t('transaction.capacity')} value={<NodeCellCapacityAmount cell={cell} />} />
       </TransactionCellCardPanel>
     )
   }
@@ -277,7 +238,7 @@ export default ({
         </div>
 
         <div className="transactionCellCapacity">
-          <TransactionCellCapacityAmount cell={cell} />
+          <NodeCellCapacityAmount cell={cell} />
         </div>
 
         <div className="transactionDetailCellInfo">
